+++ google2/CHANGELOG.md
@@ -11,7 +11,7 @@ Changed:
 Added:
 - Added `GoogleDataprocService.startCluster`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-TRAVIS-REPLACE-ME"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-4d379e3"`
 
 ## 0.18
 Added: