+++ google2/CHANGELOG.md
@@ -9,11 +9,13 @@ Changed:
 - Removed `RetryConfig` from `GoogleDataprocService` constructors
 - `GoogleComputeInterpreter` now returns none if it encounters a disabled billing project during `getInstance`
 - Added `GoogleBillingInterpreter` and `GoogleBillingService`
+- Added `createDataset` and `setDatasetIam` to `GoogleBigQueryService`
+- Added new constructor to `GoogleBigQueryService` that accepts path to credentials JSON
 
 Added:
 - Added `GoogleDataprocService.startCluster`
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-1aba7fd"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-TRAVIS-REPLACE-ME"`
 
 ## 0.18
 Added: