+++ google2/CHANGELOG.md
@@ -11,8 +11,9 @@ Changed:
 
 Added:
 - Added `GoogleDataprocService.startCluster`
-- Added `getPersistentVolumeClaims` to `KubernetesService`
+- Added `listPersistentVolumeClaims` to `KubernetesService`
 - Added `GoogleBillingInterpreter` and `GoogleBillingService`
+- Added `deletePv` to `KubernetesService`
 
 Dependency Updates:
 ```
@@ -23,7 +24,7 @@ Update http4s-blaze-client, http4s-circe, ... from 0.21.16 to 0.21.18 (#499) (2 
 Update sbt from 1.4.6 to 1.4.7 (#500) (2 minutes ago) <Scala Steward>
 ```
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-3889a04"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.19-TRAVIS-REPLACE-ME"`
 
 ## 0.18
 Added: